Move the notification lambda handler to /lambda
