+++ lambda/notifications/notifications.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"dont-slack-evil/notifications"
+
+	"github.com/aws/aws-lambda-go/lambda"
+)
+
+func main() {
+	lambda.Start(notifications.LambdaHandler)
+}
